% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubedb.R
\name{query_heatmap}
\alias{query_heatmap}
\title{query heatmap of time series}
\usage{
query_heatmap(
  tubedb,
  plot,
  sensor,
  quality = "physical",
  interpolated = FALSE,
  start = NULL,
  end = NULL,
  year = NULL,
  month = NULL,
  day = NULL,
  by_year = TRUE,
  time_scale = TRUE
)
}
\arguments{
\item{tubedb}{instance of TubeDB class}

\item{plot}{character, plot/station}

\item{sensor}{character sensor name}

\item{quality}{quality checks, one of: no, physical, step, empirical}

\item{interpolated}{apply interpolation, FALSE or TRUE}

\item{start}{start time of timeseries}

\item{end}{end time of timeseries}

\item{year}{get data from one full year only (You may use this 'year' parameter OR 'start', 'end' parameters)}

\item{month}{get data of one full month of one year only, parameter year neads to be specified (You may use this 'month' parameter OR 'start'/'end' parameters)}

\item{day}{get data of one full day of one month only, parameter month neads to be specified (You may use this 'day' parameter OR 'start'/'end' parameters)}

\item{by_year}{draw all data in one row or draw one row per year}

\item{time_scale}{draw a time scale}
}
\value{
nativeRaster
}
\description{
Query time series as heatmap pixel-image ('nativeRaster') at one plot of several sensors from TubeDB server and create timestamps of specified type.
}
\details{
Mandatory parameters: \strong{tubedb}, \strong{plot}, \strong{sensor}
}
\section{Time span of time series}{


no time parameter -> full avaiable timespan

\code{year} parameter -> one full year

\code{year} and \code{month} parameter -> one full month

\code{start} and \code{end} parameter -> exact time span

Format of \code{start} and \code{end} parameter: (character) yyyy-MM-ddTHH:mm, abbreviations allowed. e.g.  2018-12-31T23:59, 2018-12-31T23, 2018-12-31, 2018-12, 2018. '\code{end}' abbreviations are filled, e.g. 2018-01 -> 2018-01-31T23:59
}

\examples{
library(rTubeDB)
tubedb <- TubeDB(url="http://localhost:8080", user="user", password="password")
rs <- regions(tubedb)
r <- region_metadata(tubedb, "REG1")
img <- query_heatmap(tubedb, plot="PLOT1", sensor="Ta_200", start="2017", end="2019")
grid::grid.raster(img)
}
\seealso{
\link{TubeDB} \link{TubeDB} \link{query_timeseries} \link{query_regions} \link{query_region_plots} \link{query_region_sensors} \link{query_diagram}
}
\author{
woellauer
}
