% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubedb.R
\name{query_region_plot_sensors}
\alias{query_region_plot_sensors}
\title{query sensor names of one plot}
\usage{
query_region_plot_sensors(tubedb, regionID, plotID)
}
\arguments{
\item{tubedb}{instance of TubeDB class}
}
\value{
vector
}
\description{
Get sensor names of one plot of one region (project) from TubeDB as vector
}
\examples{
library(rTubeDB)
tubedb <- TubeDB(url="http://localhost:8080", user="user", password="password")
rsdf <- query_regions(tubedb)
rpdf <- query_region_plots(tubedb, "BALE")
rsrdf <- query_region_sensors(tubedb, "BALE")
rpmdf <- query_region_plot_metadata(tubedb, "BALE", "BALE001")
rpsdf <- query_region_plot_sensors(tubedb, "BALE", "BALE001")
}
\seealso{
\link{TubeDB} \link{query_timeseries} \link{query_regions_list} \link{query_regions}
}
\author{
woellauer
}
