% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubedb.R
\name{query_region_stations}
\alias{query_region_stations}
\title{query station information}
\usage{
query_region_stations(tubedb, regionID)
}
\arguments{
\item{tubedb}{instance of TubeDB class}
}
\value{
data.frame of station information
}
\description{
Get station information of one region (projects) from TubeDB as data.frame. One plot (a location) may contain several stations.
}
\examples{
library(rTubeDB)
tubedb <- TubeDB(url="http://localhost:8080", user="user", password="password")
rsl <- query_regions_list(tubedb)
rsdf <- query_regions(tubedb)
rml <- query_region_metadata(tubedb, "BALE")
ril <- query_region_info_list(tubedb, "BALE")
ridf <- query_region_info(tubedb, "BALE")
rgsl <- query_region_general_stations_list(tubedb, "BALE")
rgsdf <- query_region_general_stations(tubedb, "BALE")
rpl <- query_region_plots_list(tubedb, "BALE")
rpdf <- query_region_plots(tubedb, "BALE")
rsrl <- query_region_sensors_list(tubedb, "BALE")
rsrdf <- query_region_sensors(tubedb, "BALE")
rstl <- query_region_stations_list(tubedb, "BALE")
rstdf <- query_region_stations(tubedb, "BALE")
}
\seealso{
\link{TubeDB} \link{query_timeseries} \link{query_regions_list} \link{query_regions}
}
\author{
woellauer
}
