% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubedb.R
\name{query_regions}
\alias{query_regions}
\title{query regions}
\usage{
query_regions(tubedb)
}
\arguments{
\item{tubedb}{instance of TubeDB class}
}
\value{
data.frame of regions
}
\description{
Get regions (projects) from TubeDB as data.frame of regions
}
\examples{
library(rTubeDB)
rsl <- query_regions_list(tubedb)
rsdf <- query_regions(tubedb)
rml <- query_region_metadata(tubedb, "BALE")
ril <- query_region_info_list(tubedb, "BALE")
ridf <- query_region_info(tubedb, "BALE")
rgsl <- query_region_general_stations_list(tubedb, "BALE")
rgsdf <- query_region_general_stations(tubedb, "BALE")
rpl <- query_region_plots_list(tubedb, "BALE")
rpdf <- query_region_plots(tubedb, "BALE")
rsrl <- query_region_sensors_list(tubedb, "BALE")
rsrdf <- query_region_sensors(tubedb, "BALE")
rstl <- query_region_stations_list(tubedb, "BALE")
rstdf <- query_region_stations(tubedb, "BALE")
}
\seealso{
\link{TubeDB} \link{query_timeseries} \link{query_regions_list} \link{query_region_metadata}
}
\author{
woellauer
}
