% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubedb.R
\name{query_timeseries}
\alias{query_timeseries}
\title{query time series}
\usage{
query_timeseries(tubedb, plot, sensor, aggregation = "hour",
  quality = "physical", interpolated = FALSE, start = NULL,
  end = NULL, year = NULL, month = NULL,
  datetimeFormat = "character", colYear = FALSE, colMonth = FALSE,
  colDay = FALSE)
}
\arguments{
\item{tubedb}{instance of TubeDB class}

\item{plot}{character, plot/station}

\item{sensor}{character or list of character, sensor or list of sensors}

\item{aggregation}{time steps, one of: raw hour day week month year}

\item{quality}{quality checks, one of: no, physical, step, empirical}

\item{interpolated}{apply interpolation, FALSE or TRUE}

\item{start}{start time of timeseries}

\item{end}{end time of timeseries}

\item{year}{get data from one full year only (You may use this 'year' parameter OR 'start', 'end' parameters)}

\item{month}{get data of one full month of one year only, parameter year neads to be specified (You may use this 'month' parameter OR 'start'/'end' parameters)}

\item{datetimeFormat}{character, requested type of timestamps. one of: "character", "POSIXct", "POSIXlt"}

\item{colYear}{add numeric year column in resulting data.frame (calendar year)}

\item{colMonth}{add numeric month column in resulting data.frame (1 to 12)}

\item{colDay}{add numeric day column in resulting data.frame (1 to 31)}
}
\value{
data.frame time series with datetime column + sensor columns
}
\description{
Query time series at one plot of several sensors from TubeDB server and create timestamps of specified type.
}
\details{
Mandatory parameters: \strong{tubedb}, \strong{plot}, \strong{sensor}
}
\section{Time span of time series}{


no time parameter -> full avaiable timespan

\code{year} parameter -> one full year

\code{year} and \code{month} parameter -> one full month

\code{start} and \code{end} parameter -> exact time span

Format of \code{start} and \code{end} parameter: (character) yyyy-MM-ddTHH:mm, abbreviations allowed. e.g.  2018-12-31T23:59, 2018-12-31T23, 2018-12-31, 2018-12, 2018. '\code{end}' abbreviations are filled, e.g. 2018-01 -> 2018-01-31T23:59
}

\examples{
library(rTubeDB)
tubedb <- TubeDB(url="http://localhost:8080", user="user", password="password")
rs <- regions(tubedb)
r <- region_metadata(tubedb, "REG1")
ts <- query_timeseries(tubedb, plot="PLOT1", sensor="Ta_200", datetimeFormat="POSIXlt")
str(ts)
}
\seealso{
\link{TubeDB} \link{regions} \link{region_metadata} \link{POSIXct} \link{POSIXlt} \link{read_timeseries}
}
\author{
woellauer
}
