% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubedb.R
\name{read_timeseries}
\alias{read_timeseries}
\title{read time series from file or textConnection}
\usage{
read_timeseries(file, datetimeFormat = "character")
}
\arguments{
\item{file}{file or textConnection}

\item{datetimeFormat}{requested type of timestamps. one of: "character", "POSIXct", "POSIXlt"}
}
\value{
data.frame time series with datetime column + sensor columns
}
\description{
Read time series of TubeDB CSV-format from file or textConnection and create timestamps of specified type.
}
\seealso{
\link{POSIXct} \link{POSIXlt} \link{query_timeseries}
}
\author{
woellauer
}
