% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubedb.R
\name{region_metadata}
\alias{region_metadata}
\title{region metadata}
\usage{
region_metadata(tubedb, regionID)
}
\arguments{
\item{tubedb}{instance of TubeDB class}

\item{regionID}{ID of region}
}
\value{
metadata as list
}
\description{
Get metadata of one region from TubeDB.
}
\examples{
library(rTubeDB)
tubedb <- TubeDB(url="http://localhost:8080", user="user", password="password")
rs <- regions(tubedb)
r <- region_metadata(tubedb, "REG1")
ts <- query_timeseries(tubedb, plot="PLOT1", sensor="Ta_200", datetimeFormat="POSIXlt")
str(ts)
}
\seealso{
\link{TubeDB} \link{regions} \link{query_timeseries}
}
\author{
woellauer
}
