% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubedb.R
\name{regions}
\alias{regions}
\title{regions}
\usage{
regions(tubedb)
}
\arguments{
\item{tubedb}{instance of TubeDB class}
}
\value{
list of regions
}
\description{
Get regions (projects) from TubeDB.
}
\examples{
library(rTubeDB)
tubedb <- TubeDB(url="http://localhost:8080", user="user", password="password")
rs <- regions(tubedb)
r <- region_metadata(tubedb, "REG1")
ts <- query_timeseries(tubedb, plot="PLOT1", sensor="Ta_200", datetimeFormat="POSIXlt")
str(ts)
}
\seealso{
\link{TubeDB} \link{region_metadata} \link{query_timeseries}
}
\author{
woellauer
}
