% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubedb.R
\docType{class}
\name{TubeDB-class}
\alias{TubeDB-class}
\alias{TubeDB}
\title{An S4 class to represent a connection to TubeDB server}
\description{
Connecetion is checked at creation time of TubDB instance. So if url, user or password is not correct an error is thrown.
}
\section{Slots}{

\describe{
\item{\code{url}}{URL to TubeDB server}

\item{\code{user}}{user name of account at TubeDB server}

\item{\code{user}}{user password of account at TubeDB server}
}}

\examples{
# load package
library(rTubeDB)

# open TubeDB server connection
tubedb <- rTubeDB::TubeDB(url="http://127.0.0.1:8080", user="user", password="password")

# get regions/projects as data.frame
regionDF <- rTubeDB::query_regions(tubedb)

# get plots of one region as data.frame
plotDF <- rTubeDB::query_region_plots(tubedb, "BALE")

# get all sensors of all plots of one region as data.frame
sensorDF <- rTubeDB::query_region_sensors(tubedb, "BALE")

# get climate time series of two sensors over two plots with default processing at full time span as data.frame
tsDF <- rTubeDB::query_timeseries(tubedb, plot=c("BALE001", "BALE002"), sensor=c("Ta_200", "rH_200"), datetimeFormat="POSIXlt")

# show time series
plot(tsDF$datetime, tsDF$Ta_200)
}
\seealso{
\link{query_timeseries} \link{query_regions} \link{query_region_plots} \link{query_region_sensors}
}
\author{
woellauer
}
