% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubedb.R
\docType{methods}
\name{TubeDB}
\alias{TubeDB}
\title{constructor for \link{TubeDB-class}}
\usage{
\S4method{initialize}{TubeDB}(.Object, url, user, password)
}
\description{
This is the constructor.
}
