% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubedb.R
\name{write_timeseries}
\alias{write_timeseries}
\title{write time series  to file or textConnection}
\usage{
write_timeseries(x, file, sensors = NULL)
}
\arguments{
\item{x}{time series as data.frame}

\item{file}{file or textConnection}

\item{sensors}{sensor columns}
}
\description{
Write time series data.frame to file or textConnection in TubeDB CSV-format.
}
\details{
datetime cloumn need to exist. Resulting CSV-file contains as first row datetime and following columns with sensors.
If sensors parameter is missing all columns are included, exept 'year', 'month', 'day' (maybe created by read_timeseries).
}
\seealso{
\link{read_timeseries} \link{query_timeseries} \link{TubeDB} \link{query_regions} \link{query_region_plots} \link{query_region_sensors}
}
\author{
woellauer
}
