% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubedb.R
\name{apiGet}
\alias{apiGet}
\title{direct TubeDB API request}
\usage{
apiGet(tubedb, method, param_list = NULL, processed = TRUE)
}
\arguments{
\item{tubedb}{instance of TubeDB class}

\item{method}{API function}

\item{param_list}{list of query parameters}

\item{processed}{response is converted to appropriate objects or raw respone is returned}
}
\value{
response
}
\description{
Query time series from TubeDB server and create timestamps of specified type. If query could not be processed an error is throwen.
}
\seealso{
\link{TubeDB} \link{regions} \link{region_metadata} \link{query_timeseries}
}
\author{
woellauer
}
