% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubedb.R
\name{query_timeseries}
\alias{query_timeseries}
\title{query time series}
\usage{
query_timeseries(tubedb, plot, sensor, aggregation = "hour",
  quality = "physical", interpolated = FALSE, year = NULL,
  month = NULL, datetimeFormat = "character")
}
\arguments{
\item{tubedb}{instance of TubeDB class}

\item{plot}{character, plot/station}

\item{sensor}{character or list of character, sensor or list of sensors}

\item{aggregation}{time steps, one of: raw hour day week month year}

\item{quality}{quality checks, one of: no, physical, step, empirical}

\item{interpolated}{apply interpolation, FALSE or TRUE}

\item{year}{get data from one year only}

\item{month}{get data of one month of one year only, year neads to be specified}

\item{datetimeFormat}{character, requested type of timestamps. one of: "character", "POSIXct", "POSIXlt"}
}
\value{
data.frame time series with datetime column + sensor columns
}
\description{
Query time series from TubeDB server and create timestamps of specified type.
}
\examples{
library(rTubeDB)
tubedb <- TubeDB(url="http://localhost:8080", user="user", password="password")
rs <- regions(tubedb)
r <- region_metadata(tubedb, "REG1")
ts <- query_timeseries(tubedb, plot="PLOT1", sensor="Ta_200", datetimeFormat="POSIXlt")
str(ts)
}
\seealso{
\link{TubeDB} \link{regions} \link{region_metadata} \link{POSIXct} \link{POSIXlt} \link{read_timeseries}
}
\author{
woellauer
}
