% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubedb.R
\name{read_timeseries}
\alias{read_timeseries}
\title{read time series from file or textConnection}
\usage{
read_timeseries(file, datetimeFormat = "character", colYear = FALSE,
  colMonth = FALSE, colDay = FALSE, colHour = FALSE)
}
\arguments{
\item{file}{file or textConnection}

\item{datetimeFormat}{requested type of timestamps. one of: "character", "POSIXct", "POSIXlt"}

\item{colYear}{add numeric year column in resulting data.frame (calendar year)}

\item{colMonth}{add numeric month column in resulting data.frame (1 to 12)}

\item{colDay}{add numeric day column in resulting data.frame (1 to 31)}

\item{colHour}{add numeric hour column in resulting data.frame (0 to 23)}
}
\value{
data.frame time series with datetime column + sensor columns
}
\description{
Read time series of TubeDB CSV-format from file or textConnection and create timestamps of specified type.
}
\seealso{
\link{POSIXct} \link{POSIXlt} \link{query_timeseries} \link{TubeDB} \link{query_regions} \link{query_region_plots} \link{query_region_sensors}
}
\author{
woellauer
}
