% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tubedb.R
\docType{class}
\name{TubeDB-class}
\alias{TubeDB-class}
\alias{TubeDB}
\title{An S4 class to represent a connection to TubeDB server}
\description{
Connecetion is checked at creation time of TubDB instance. So if url, user or password is not correct an error is thrown.
}
\section{Slots}{

\describe{
\item{\code{url}}{URL to TubeDB server}

\item{\code{user}}{user name of account at TubeDB server}

\item{\code{user}}{user password of account at TubeDB server}
}}

\examples{
library(rTubeDB)
tubedb <- TubeDB(url="http://localhost:8080", user="user", password="password")
rs <- regions(tubedb)
r <- region_metadata(tubedb, "REG1")
ts <- query_timeseries(tubedb, plot="PLOT1", sensor="Ta_200", datetimeFormat="POSIXlt")
str(ts)
}
\seealso{
\link{regions} \link{region_metadata} \link{query_timeseries} \link{apiGet}
}
\author{
woellauer
}
